requirements.txt
alabaster==0.7.12
astroid==2.9.3
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
docutils==0.17.1
h11==0.13.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.1
isort==5.10.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
markdown-it-py==2.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
myst-parser==0.17.0
outcome==1.1.0
packaging==21.3
platformdirs==2.5.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pylint==2.12.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.1.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.1.1
urllib3==1.26.8
webdriver-manager==3.5.3
wrapt==1.13.3
wsproto==1.0.0
zipp==3.7.0
